Refactored routes
- Moved user-server routes to /admin/routes/api/user-server
